+++ now.json
@@ -28,7 +28,7 @@
     {
       "continue": true,
       "headers": {
-        "cache-control": "max-age=600, s-maxage=60"
+        "cache-control": "max-age=600, s-maxage=120"
       },
       "src": ".*"
     },
@@ -48,13 +48,13 @@
     {
       "continue": true,
       "headers": {
-        "content-security-policy": "connect-src 'self' https://*.algolia.net https://*.algolianet.com https://fonts.googleapis.com https://fonts.gstatic.com; default-src 'none'; font-src https://fonts.gstatic.com; img-src 'self' data: https://i.ytimg.com; manifest-src 'self'; script-src 'self' https://storage.googleapis.com; style-src 'self' 'unsafe-inline' https://fonts.googleapis.com; worker-src 'self'"
+        "content-security-policy": "connect-src 'self' https://*.algolia.net https://*.algolianet.com https://fonts.googleapis.com https://fonts.gstatic.com; default-src 'none'; font-src https://fonts.gstatic.com; img-src 'self' data: https://i.ytimg.com; manifest-src 'self'; script-src 'self' 'unsafe-inline' https://storage.googleapis.com; style-src 'self' 'unsafe-inline' https://fonts.googleapis.com; worker-src 'self'"
       },
       "src": ".*"
     },
     {
       "headers": {
-        "cache-control": "max-age=0"
+        "cache-control": "max-age=300,s-maxage=60"
       },
       "src": "/(?:search)?"
     },
