--- src/components/molecules/search-results/index.ts
@@ -1,1 +1,0 @@
-export { default } from './search-results'
