+++ src/pages/search.tsx
@@ -1,46 +1,14 @@
 import { NextPage } from 'next'
-import React, { ReactElement, useCallback, useState, useEffect } from 'react'
+import React, { ReactElement } from 'react'
 import { Helmet } from 'react-helmet'
-import { InView } from 'react-intersection-observer'
-import Spinner from '../components/atoms/spinner'
-import SearchResults from '../components/molecules/search-results'
-import search from '../lib/search'
+import Search from '../components/molecules/search'
 import { getTitle } from '../lib/title'
-import Video from '../types/video'
 
 export type SearchProps = {
-  hasNext: boolean
-  hits: Video[]
   query: string
 }
 
-const Search: NextPage<SearchProps> = ({
-  hasNext: initialHasNext,
-  hits,
-  query
-}): ReactElement => {
-  const [results, setResults] = useState<Video[]>(hits)
-  const [page, setPage] = useState<number>(0)
-  const [hasNext, setHasNext] = useState<boolean>(initialHasNext)
-
-  const handleChange = useCallback((): void => {
-    search<Video>(query, { page: page + 1 })
-      .then(({ hits, nbPages, page: actualPage }): void => {
-        setHasNext(actualPage < nbPages - 1)
-        setPage(actualPage)
-        setResults((prevResults): Video[] => prevResults.concat(hits))
-      })
-      .catch((): void => {
-        setHasNext(false)
-      })
-  }, [page, query])
-
-  useEffect((): void => {
-    setResults(hits)
-    setPage(0)
-    setHasNext(initialHasNext)
-  }, [hits, initialHasNext])
-
+const SearchPage: NextPage<SearchProps> = ({ query }): ReactElement => {
   const title = [query, getTitle()].filter(Boolean).join(' - ')
   const description = process.env.ANIMARE_SEARCH_DESCRIPTION
   const baseUrl = process.env.ANIMARE_SEARCH_BASE_URL || 'https://example.com'
@@ -65,43 +33,19 @@ const Search: NextPage<SearchProps> = ({
         <meta content="summary_large_image" name="twitter:card" />
       </Helmet>
 
-      <SearchResults values={results} />
-
-      <footer className="search__footer">
-        {hasNext && (
-          <InView as="div" className="search__loading" onChange={handleChange}>
-            <Spinner aria-label="読み込み中..." />
-          </InView>
-        )}
-      </footer>
-
-      <style jsx>{`
-        .search__footer {
-          padding: 2rem 0;
-        }
-
-        :global(.search__loading) {
-          align-items: center;
-          display: flex;
-          height: 100%;
-          justify-content: center;
-        }
-      `}</style>
+      <Search query={query} />
     </>
   )
 }
 
-Search.getInitialProps = async ({ query }): Promise<SearchProps> => {
+SearchPage.getInitialProps = async ({ query }): Promise<SearchProps> => {
   const q = (Array.isArray(query.q) ? query.q[0] : query.q) || ''
-  const { hits, nbPages } = await search<Video>(q)
 
   if (typeof window !== 'undefined') window.scrollTo(0, 0)
 
   return {
-    hasNext: nbPages > 1,
-    hits,
     query: q
   }
 }
 
-export default Search
+export default SearchPage
